Rename depversions to clicompat and improve code quality
- Rename libs/depversions/ to libs/clicompat/ (package + files)
- Rename internal/build/dep_versions.go to clicompat.go
- Rename EmbeddedManifestJSON to CLICompatManifestJSON
- Extract devVersionPrefix const for "0.0.0-dev"
- Wrap both errors with %w in FetchManifest fallback
- Add template-v tag validation to bump-cli-compat skill
- Remove confusing positional args from skill (named flags only)
- Fix README: "After each AppKit or Agent Skills release"
