+++ .github/OWNERS
@@ -61,7 +61,7 @@
 
 # CLI compatibility manifest
 /internal/build/cli-compat.json  team:eng-apps-devex team:platform
-/libs/depversions/               team:eng-apps-devex team:platform
+/libs/clicompat/                 team:eng-apps-devex team:platform
 
 # Experimental
 /experimental/aitools/      team:eng-apps-devex @lennartkats-db
